backend/requirements.txt
python-dotenv>=1.0
django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
google-auth>=2.36
requests>=2.32
gunicorn>=23.0
psycopg[binary]>=3.2
dj-database-url>=2.3
whitenoise>=6.8
django-cors-headers>=4.6
django-storages[s3]>=1.14
Pillow>=11.0
django-imagekit>=5.0
